+++ basic/src/main/java/com/dede/basic/LifecycleExt.kt
@@ -1,4 +1,5 @@
 @file:Suppress("DEPRECATION")
+@file:JvmName("LifecycleUtils")
 
 package com.dede.basic